requirements.txt
# /requirements.txt
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==2.1.5
six==1.16.0
Werkzeug==3.0.3
fastapi
uvicorn


python-dotenv
python-jose
six
google-cloud-secret-manager
google-cloud-logging
google-cloud-pubsub
loguru
pyyaml~=6.0.1
flasgger~=0.9.7.1
python-json-logger
google-cloud-logging==3.10.0
google-cloud-logging==3.10.0
grpcio
grpcio-tools


utils~=1.0.2
